STORAGE2: Don't copy files that don't exist

+++ src/storage2_node.cpp
@@ -801,6 +801,10 @@ struct Storage2Node::Impl {
       config_output << config;
     }
     for(auto& file : files) {
+      if(!std::filesystem::exists(file)) {
+        THALAMUS_LOG(error) << "File doesn't exist: " << file.string();
+        continue;
+      }
       std::filesystem::copy_file(file, filepath.parent_path() / (
             file.stem().filename().string()
             + filepath.stem().extension().string()
